+++ scripts/install-dependency-packages.sh
@@ -28,7 +28,3 @@ apt-get update
 # C modules, particularly when upgrading to newer Python versions.  libffi-dev
 # is sometimes needed to build cffi (a cryptography dependency).
 apt-get -y install --no-install-recommends build-essential libffi-dev
-
-# Delete cached files we don't need anymore:
-apt-get clean
-rm -rf /var/lib/apt/lists/*