+++ src/hoverdrive/config.py
@@ -22,9 +22,7 @@ class Config(BaseSettings):
 
     name: str = Field("hoverdrive", title="Name of application")
 
-    path_prefix: str = Field(
-        "/hoverdrive", title="URL prefix for application"
-    )
+    path_prefix: str = Field("/hoverdrive", title="URL prefix for application")
 
     profile: Profile = Field(
         Profile.development, title="Application logging profile"
@@ -34,6 +32,11 @@ class Config(BaseSettings):
         None,
         title="Slack webhook for alerts",
         description="If set, alerts will be posted to this Slack webhook",
+    )
+
+    ook_url: str = Field(
+        "https://roundtable.lsst.cloud/ook",
+        description="Base URL for the Ook API",
     )