+++ tests/handlers/external_test.py
@@ -3,7 +3,8 @@
 from __future__ import annotations
 
 import pytest
-from httpx import AsyncClient
+import respx
+from httpx import AsyncClient, Response
 
 from hoverdrive.config import config
 
@@ -20,3 +21,193 @@ async def test_get_index(client: AsyncClient) -> None:
     assert isinstance(metadata["description"], str)
     assert isinstance(metadata["repository_url"], str)
     assert isinstance(metadata["documentation_url"], str)
+
+
+@pytest.mark.asyncio
+async def test_get_column_docs_link_redirect(
+    client: AsyncClient, respx_mock: respx.Router
+) -> None:
+    """Test ``GET /hoverdrive/column-docs-links``."""
+    respx_mock.get(
+        "https://roundtable.lsst.cloud/ook/links/domains/sdm/schemas/"
+        "dp02_dc2_catalogs/tables/Object/columns/detect_isPrimary"
+    ).mock(
+        return_value=Response(
+            status_code=200,
+            json=[
+                {
+                    "url": (
+                        "https://sdm-schemas.lsst.io/dp02.html#Object.detect_isPrimary"
+                    ),
+                    "type": "schema_browser",
+                    "title": "dp02_dc2_catalogs.Object.detect_isPrimary",
+                    "collection_title": "SDM Schema Browser",
+                }
+            ],
+        )
+    )
+    response = await client.get(
+        "/hoverdrive/column-docs-links",
+        params={
+            "table": "dp02_dc2_catalogs.Object",
+            "column": "detect_isPrimary",
+            "redirect": True,
+        },
+    )
+    assert response.status_code == 307
+    assert response.headers["Location"] == (
+        "https://sdm-schemas.lsst.io/dp02.html#Object.detect_isPrimary"
+    )
+
+
+@pytest.mark.asyncio
+async def test_get_table_docs_link_redirect(
+    client: AsyncClient, respx_mock: respx.Router
+) -> None:
+    """Test ``GET /hoverdrive/table-docs-links``."""
+    respx_mock.get(
+        "https://roundtable.lsst.cloud/ook/links/domains/sdm/schemas/"
+        "dp02_dc2_catalogs/tables/Object"
+    ).mock(
+        return_value=Response(
+            status_code=200,
+            json=[
+                {
+                    "url": ("https://sdm-schemas.lsst.io/dp02.html#Object"),
+                    "type": "schema_browser",
+                    "title": "dp02_dc2_catalogs.Object",
+                    "collection_title": "SDM Schema Browser",
+                }
+            ],
+        )
+    )
+    response = await client.get(
+        "/hoverdrive/table-docs-links",
+        params={
+            "table": "dp02_dc2_catalogs.Object",
+            "redirect": True,
+        },
+    )
+    assert response.status_code == 307
+    assert response.headers["Location"] == (
+        "https://sdm-schemas.lsst.io/dp02.html#Object"
+    )
+
+
+@pytest.mark.asyncio
+async def test_get_column_docs_links_redirect_multiple_columns(
+    client: AsyncClient, respx_mock: respx.Router
+) -> None:
+    """Test ``GET /hoverdrive/column-docs-links`` with multiple columns for
+    redirection, which is not supported.
+    """
+    response = await client.get(
+        "/hoverdrive/column-docs-links",
+        params={
+            "table": "dp02_dc2_catalogs.Object",
+            "column": [
+                "detect_isPrimary",
+                "detect_isPrimary_2",
+            ],
+            "redirect": True,
+        },
+    )
+    assert response.status_code == 400
+    assert response.json() == {
+        "detail": [
+            {
+                "loc": ["query", "column"],
+                "msg": ("Only one column name is supported for redirection."),
+                "type": "bad_link_redirect_request",
+            }
+        ]
+    }
+
+
+@pytest.mark.asyncio
+async def test_get_table_docs_links_redirect_multiple_tables(
+    client: AsyncClient, respx_mock: respx.Router
+) -> None:
+    """Test ``GET /hoverdrive/table-docs-links`` with multiple tables for a
+    redirect, which is not supported.
+    """
+    response = await client.get(
+        "/hoverdrive/table-docs-links",
+        params={
+            "table": ["dp02_dc2_catalogs.Object", "dp02_dc2_catalogs.Visit"],
+            "redirect": True,
+        },
+    )
+    assert response.status_code == 400
+    assert response.json() == {
+        "detail": [
+            {
+                "loc": ["query", "table"],
+                "msg": ("Only one table name is supported for redirection."),
+                "type": "bad_link_redirect_request",
+            }
+        ]
+    }
+
+
+@pytest.mark.asyncio
+async def test_get_column_docs_links_redirect_required(
+    client: AsyncClient, respx_mock: respx.Router
+) -> None:
+    """Test ``GET /hoverdrive/column-docs-links`` that redirect is required."""
+    respx_mock.get(
+        "https://roundtable.lsst.cloud/ook/links/domains/sdm/schemas/"
+        "dp02_dc2_catalogs/tables/Object/columns/detect_isPrimary"
+    ).mock(
+        return_value=Response(
+            status_code=200,
+            json=[
+                {
+                    "url": (
+                        "https://sdm-schemas.lsst.io/dp02.html#Object.detect_isPrimary"
+                    ),
+                    "type": "schema_browser",
+                    "title": "dp02_dc2_catalogs.Object.detect_isPrimary",
+                    "collection_title": "SDM Schema Browser",
+                }
+            ],
+        )
+    )
+    response = await client.get(
+        "/hoverdrive/column-docs-links",
+        params={
+            "table": "dp02_dc2_catalogs.Object",
+            "column": "detect_isPrimary",
+        },
+    )
+    assert response.status_code == 501
+
+
+@pytest.mark.asyncio
+async def test_get_table_docs_links_redirect_required(
+    client: AsyncClient, respx_mock: respx.Router
+) -> None:
+    """Test ``GET /hoverdrive/table-docs-links`` that redirect is required."""
+    respx_mock.get(
+        "https://roundtable.lsst.cloud/ook/links/domains/sdm/schemas/"
+        "dp02_dc2_catalogs/tables/Object"
+    ).mock(
+        return_value=Response(
+            status_code=200,
+            json=[
+                {
+                    "url": ("https://sdm-schemas.lsst.io/dp02.html#Object"),
+                    "type": "schema_browser",
+                    "title": "dp02_dc2_catalogs.Object",
+                    "collection_title": "SDM Schema Browser",
+                }
+            ],
+        )
+    )
+    response = await client.get(
+        "/hoverdrive/table-docs-links",
+        params={
+            "table": "dp02_dc2_catalogs.Object",
+        },
+    )
+    assert response.status_code == 501